+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Create "rightClick" method for browser tabs.
 * Create "evaluateOnNewDocument" method for browser tabs.
 
 1.0.0 - 09-05-2021